Modified the Chart Query to fill chart Exmple
Noticethe comments that explain the field names and how to assign them. Note the you must go find out what a Anonymous class is, the is explaination under the that favoriate web site of mine that expalins the Entity Framework - look on hte menu on the left on that web site.